Use lazy desrialization for SQL data in transit (#20402)
Before we used `Object[]` as the data type used for data sent between vertices
of a SQL job. Values in the row were always deserialized, which created
unnecessary overhead if e.g. only one field from the row was used (is some cases
even no field is accessed).

This PR replaces the `Object[]` with `JetSqlRow` which lazily deserializes
fields as they are accessed.