Remove instance-tracking config from config template
Instance tracking is disabled by default for normal build, but for NLC
build it is enabled. Having it in the default config template (used in
zip, Docker and elsewhere) disables it also for the NLC build.

Removing the config from the template is simplest fix, compared to
creating another configuration file for the NLC build.

This is rather exotic feature, so we don't need this snippet in the
default configs, however it is still present in the full example
configs.

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/4349