Fix hanging HzStartTest
The Hazelcast process wasn't being killed correctly and in combination
with java.lang.ProcessBuilder#inheritIO the JVM running the test was
hanging at shut down (the test passed correctly).

There were 2 issues causing improper shutdown:

 - the scripts hz/hz-start were not using exec command which replaces
current shell process with the command, this resulted in killing the
shell, but that doesn't immediatelly kill the HazelcastServerCommandLine
process.  Using exec means we kill the HazelcastServerCommandLine
directly.

 - similarly the HazelcastServerCommandLine starts a new process
(HazelcastMemberStarter) and when it is killed it doesn't automatically
kill the new process. Added shutdown hook so the process is killed in
normal scenario (it won't be killed correclty if `kill -9` is used, but
we can't do anything about that).

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/4425