Fix SqlErrorTest.testMemberLeave
I suppose that if the shutdown happens at a very early moment, the job fails
with HazelcastInstanceNotActiveException instead of MemberLeftException. This
very behavior is tested in `testMemberLeaveDuringQueryAfterImmediateShutdown()`.

The fix is to not just for the job te be in RUNNING state, but wait until at
least one row is received from the query.

Fixes #20501