Fix SqlErrorTest.testTimeout failure (#20514)
The test used a batch source and a timeout of 1ms. It could happen that the
query completes under 1ms, in which case the test fails. It's easy to reproduce
in the original code by increasing the timeout.

The fix is to use a streaming source. I also removed the `testTimeout_fetch()`
method because it's not simple to implement and there's no difference in code -
it's communicated to the caller when iterating the SqlResult on the member, and
that's the same code handling both `execute` and `fetch` client messages.

Fixes #20011