Fix iterator visibility (#20506)
The reason of setting iterator again is to make internal state of the
iterator also visible. Making only reference volatile is not sufficient
in this case. By re-setting iterator, we're doing a volatile write to
make all state before it visible for other threads.

Other changes are refactoring to share common functionality. Goal
of this refactoring is not to repeat same fix in multiple places.