Remove ParallelJVMTest from JobSubmissionSlownessRegressionTest
The test submits as many jobs as it can during a period of time. If
other CPU intensive tests run at the same time this will inevitably
reduce the rate, causing the test to fail.

It is still possible that other jobs on the machine may affect the test,
but it is less likely.

Fixes #19658