Fix SqlErrorTest.testCancel
It was a test issue. The SqlResult is automatically closed when the backing job
completes (for member cursors) or when all rows are fetched (for client cursor).
A sleep inserted before `res.close()` reproduced the issue (only for member,
because for client the fetch size was set to 1 and the client would never fetch
all rows at that point.

Fixed by replacing the query with a streaming one.

Fixes #20336