Send compact schemas urgently during cluster changes
We send the compact schemas we have on the client to the new cluster
we connect while the client is in the `CONNECTED_TO_CLUSTER` state.

During that state, normal invocations are not allowed, as the client
has not finished sending its state to the cluster.

So, in the remote operations we do during that state, we have to
invoke them as urgently, otherwise they won't be allowed to run.

However, we were not doing that in the `ClientSchemaService#sendAllSchemas`
method, which is part of the data the client must send to the cluster.

This PR fixes that problem.

We had a test(`testClusterRestart`) where we were checking the cluster restart
case, but the test was using a different client after the restart. The first
client was desperately retrying sending state to the cluster but failing to
do so.

I have also fixed the test by using the same client instance.