Fix thread leak that kept starting new instances (#20566)
The tests in the `MapScanMigrationStressTest` used to start a `MutatorThread`,
which was adding and removing members to the cluster. Normally, this thread was
terminated. But if the test failed, the termination code didn't execute and the
thread kept running and caused unrelated tests to fail with `Instances haven't
been shut down`.

The fix is to terminate the thread in `@After`.

Fixes #20182
Fixes #19987
Fixes #19884 
Fixes #19638

The original issue triggering this mayhem is reported in #19885, but this PR
doesn't fix that.