Add test and warning for duplicate Kafka topics (#20565)
Relates to #19948

The issue seems to be fixed by
https://github.com/hazelcast/hazelcast-jet/pull/2732. Before this commit
this test reliably fails. After it I couldn't reproduce it with
various different partition counts etc.

Adding the test to catch any future regressions and a log warning that
duplicate topics were specified. This usually means mis-configuration.

* Fix condition in assertTrueAllTheTime

The original condition means that for assertTrueAllTheTime(.., 1) the
method will
- call the assert task once
- wait 1s
- exit

After the change the assert task will be done one more time and there
will be no waiting before exit.

Other methods in the HazelcastTestSupport suffer from the same issue,
but the waiting is in order of milliseconds.