Optimize retainAndGetNamespaces() (#19462) (#20612)
During unrelated profiling, this method seems particularly slow. We used
`LinkedList` for a collection which was later used as an argument to
`HashMap.retainAll()`. The implementation iterates the map and executes
`contains()` on the `LinkedList` for each map entry. I don't know the
runtime size of the collection, but it's slow. This PR replaces the
`LinkedList` with a `HashSet`.

(cherry picked from commit bcfd51f52592f6fadf939b5d933ce9a0c1fef3c3)

Co-authored-by: Viliam Durina <viliam-durina@users.noreply.github.com>