Align HazelcastProperties#get with Properties#getProperty
When a HazelcastProperties object is initialized with a compromised
Properties object (ie one containing keys/values other than Strings),
calling HazelcastProperties#get(String) may result in a
ClassCastException (when value is not a String).
This PR aligns behaviour of HazelcastProperties#get with Properties#get:
when a non-String value is mapped to given key, then return null
instead of throwing a ClassCastException.