Remove `getRegistries` API from the CompactSerializationConfig
After an internal discussion, we have decided that it is not
useful to have a getter for the registered serializers in the
compact serialization config.

Therefore, this PR removes that API and replaces its usages
with the accessor class, we have for it.

Also, this PR replaces the word `registry` with `registration`
on some places.