Revert ExplicitTable to SQL standard
In #19886 we changed the table-function arguments to what the standard
specifies:
```
SELECT *
FROM TABLE(HOP(TABLE(foo_table), ...))
```

That is, the `foo_table` is in parentheses. However, Calcite uses
`ExplicitTable`, which is another standardized construct, and it's a
prefix-style operator, i.e.
```
SELECT *
FROM TABLE(HOP(TABLE foo_table, ...))
```
That is, the `foo_table` is separated by a space after `TABLE` keyword.

We did it by copying the `Parser.jj` template into our code and modifying.
However, this broke another standard behavior, where e.g. the query `(table t)`
behaves the same as `SELECT * FROM t` - previously, `(table(t))` was also a
valid query, which was non-standard.

This PR reverts the change back. Now we'll not support the table-function syntax
with parentheses, but we think the syntax with prefix-style operator is also
valid. We might add the function-style syntax in the future, ideally by adding
support in Calcite.