Fix termination on member leave
This was possible to happen:

- a job was submitted

- before the execution started on some member, a member left the cluster

- in `JobExecutionService.terminateExecution0()` called from
`JobExecutionService.onMemberRemoved()` we determined that the execution didn't
yet begin, so we completed the execution with `CancellationException`

- the master received the `CancellationException` and cancelled the job, while
after a member leaving, it should have restarted it.

The fix is to ensure that the execution is completed with `MemberLeftException`
instead of `CancellationException` - the master considers this as a topology
exception and correctly restarts the job. This exception also would be thrown if
the execution was already begun.

Fixes #19961