Improve logging customization [HZ-772] (#20528)
* Improve logging customization [HZ-772]

- Provide ability to customize the default pattern via LOGGING_PATTERN
env variable. To support patterns with exception inside the pattern the
default config file now sets alwaysWriteExceptions=false and `%eXe` was
added to the default pattern configuration.

- Provide ability to specify the configuration file via LOGGING_CONFIG
env variable. This takes either relative or absolute path, or file in
config directory. If not set config/log4j2.properties is used.

- Add support for JsonTemplateLayout (log4j-layout-template-json
dependency) and a built in configuration file using JsonLayout.json as
default for the template.

- The json template can be customized via LOGGING_JSON_TEMPLATE, see
https://logging.apache.org/log4j/2.x/manual/json-template-layout.html#event-templates
for out of the box available templates.

- Small cleanup related to default LOGGING_LEVEL - moved the default out
of the startup scripts