Throw exception during scan if conflict (#20579)
During scan if we detect a conflict, we throw an exception. Note that
this doesn't check other members but only the current member. To do a
consistency check cluster-wide prior to dynamic changes discussed and
won't be implemented. Basically if we have inconsistent cluster it's ok
to behave in an unexpected manner. User should start the dynamic change
when the configurations between members are in consistent state.