Do not try to serialize unsupported field types in the reflective compact serializer
We have decided to remove `char`, `Character`, `char[]`, and
`Character[]` type support from the `CompactReader/Writer` APIs,
as we thought this types are not a good fit for some non-Java clients
and most of other serialization libraries do not support it.

The users still can serialize these types as they wish (possibly as
int), but that responsibility is left to the user.

We could make an exception for the zero-config use case, as it
will be primarily used by the Java-only users, but we decided
to postpone this decision to 5.2.

Therefore, to not fail with counterintuitive errors when confronted
with these types, we decided to fail fast and throw an error which
describes what to do, in case the user wants to use these types
in their classes.