[HZ-907] Protect MulticastDiscoveryStrategy against Java deserialization attacks (#20728)
The 'safe-serialization' property is added to the MulticastDiscoveryStrategy configuration. It determines if Java Serialization is used (false) or rather the safer and portable one (true). The Java native serialization format which is used by default is sensitive to [insecure deserialization attacks](https://owasp.org/www-project-top-ten/2017/A8_2017-Insecure_Deserialization).

Safe serialization format of the MulticastMemberInfo:

```
 boolean (1b):              flag if memberInfo provided (false mean memberInfo is null)
 UTF-8   (variable length): host
 int:    (4b):              port
```

**UTF-8 format notes (see java.io.DataOutputStream.writeUTF(String)):**
Two bytes (short) provides subsequent length to read. This value is the number of bytes actually, not the length of the string. Modified UTF-8 is used for decoding.