Dont allow empty memberlist on cluster id changes
When cluster id changed we were resetting member list
with an empty memberlist. It turned out that there are implementations
that assumes it will never be empty.

Following test failure is an example of it.
https://github.com/hazelcast/hazelcast/issues/20264

Instead of clearing the memberlist on cluster id change, we
are keeping the cluster id as part of the latest snapshot
to be able to fire correct events.

Note that clearing memberlist on cluster id change was
introduced as a fix to https://github.com/hazelcast/hazelcast/pull/18245
So it is important not to break related ClientHotRestartTest

fixes https://github.com/hazelcast/hazelcast/issues/20264