Prevent infinite loop when local partition missing (#20894)
This could have happened:
- reading of local partition throws MissingPartitionException
- processor looks at partition table, the partition is assigned to the same owner
- the new split is immediately retried. Since it's local partition, the
operation is run in-thread.

The above loops forever because of #20876. But even if it is fixed, the new
algorithm will be more robust as it will not spin until the situation resolves,
but will rather do it after a delay. We only insert the delay if there was no
change in the splits. If there was, it makes sense to retry immediately.