Fix race in SingleProtocolEncoder [HZ-1008] (#20991)
SingleProtocolEncoder#signalProtocolVerified was called from the
decoder side, which runs on the input IO-thread, while
SingleProtocolEncoder#onWrite is called from the output IO-thread.
Since the shared state variables that these methods access are not
volatile, there was a data race. Also, `isDecoderReceivedProtocol` is
the first checked state variable in onWrite method, so we have to
update it last on the signal methods -> changed the order of updates to
satisfy this.

(cherry picked from commit 9e1ec99aaa0f02d6da606d8940aa3dbc98e131b8)