Remove failing compatibility test
The removed test relies on the compatibility test infrastructure
to proxy an `IMap` instance across classloaders from current to
4.0.3 version. However, the compatibility test infrastructure is
tweaked to handle Hazelcast 3 <--> 5 proxying, so won't work for
4 <--> 5 proxying.
The particular failure is in IMap interface proxying. IMap in HZ 3
is in package com.hazelcast.core while in HZ 4 & 5 it is in package
com.hazelcast.map. Since compat tests expect to be doing proxying
for 3 <--> 5 compatibility tests, a lookup for com.hazelcast.core.IMap
results in ClassNotFoundException in the test.