Fix validation of unsupported features (#20997)
We used to run UnsupportedOperationVisitor after the validation. But it could
have happened that an unsupported feature was used (e.g. UPSERT) and the error
thrown was misleading because it was thrown during the validation, such as
"Number of INSERT target columns (6) does not equal number of source items (4)",
which was confusing to the user, because the UPSERT isn't supported in the first
place. In this specific case, the target indeed had 4 columns, but UPSERT also
includes the hidden columns `__key` and `this`.