Fix failure in TestCacheManager test (#21018)
The TTL was set to 1, ttl has resolution in seconds.
When the entry is stored at e.g. 5.99 seconds and fetched at 6.01
seconds it is already expired. Increase ttl to 2 s ensures at least 1
second needs to pass between the 2 calls for the entry to expire.

Also improve test isolation of test in the same class, which made this
test failure more likely to occur.

Fixes #19527