Don't use stale information in View for querying (#21028)
We store field list of a view in the catalog. We don't strictly track
dependencies, so if a view depends on some other object (a view or a mapping),
the user is allowed to change that object. It should in turn affect the
dependent view. However, since we stored the field list and we used that list
for building schema for query validation, we used stale information, which led
to failures or incorrect results.

The fix in this PR is as follows:

- we still store the field list in catalog, but only for the purpose of
  `information_schema` queries. The stale information can still remain in these
  tables

- we don't use the stored fields for validation, rather, we get the view query
  and parse it anew.

- in order to avoid parsing of all of the views for each query optimization, we
  made the view field list lazy - it's only parsed for views actually used in
  the query

- we make sure that the view is parsed only once - if we parse it during
  validation, the result is later used in the sql-to-rel conversion.

Fixes #20032