Improve PhoneHomeCPSubsystemTest [5.1.z]
A CP subsystem operation returns when it is commited on majority of the
nodes, 2 out of 3 in this test. It can happen that an operation is not
finished on the first member (127.0.0.0:5701) and the phoneHome
collection don't see the atomic value, semaphore, .. yet.

It can be reproduced by slowing down the operation on the first member
e.g. in `com.hazelcast.cp.internal.NodeEngineRaftIntegration#runOperation`

```
if (operation instanceof GetAndSetOp
                && nodeEngine.getLocalMember().getAddress().toString().equals("[127.0.0.1]:5701")) {
            try {
                Thread.sleep(1000);
            } catch (InterruptedException e) {
                e.printStackTrace();
            }
        }
```

The test is fixed by moving the phoneHome collection and the assert to
`assertTrueEventually` block.

Fixes #21014
Fixes #20903
HZ-1038, HZ-1039

Backport of #21084