Don't allow empty memberlist on cluster id changes [API-1215] (#20818)
* Dont allow empty memberlist on cluster id changes

When cluster id changed we were resetting member list
with an empty memberlist. It turned out that there are implementations
that assumes it will never be empty.

Following test failure is an example of it.
https://github.com/hazelcast/hazelcast/issues/20264

Instead of clearing the memberlist on cluster id change, we
are keeping the cluster id as part of the latest snapshot
to be able to fire correct events.

Note that clearing memberlist on cluster id change was
introduced as a fix to https://github.com/hazelcast/hazelcast/pull/18245
So it is important not to break related ClientHotRestartTest

fixes https://github.com/hazelcast/hazelcast/issues/20264

* Refactor ClusterService to make it unittestable

Move getClient API out to reduce dependency to
HazelcastInstanceImpl

* Refactor: ClusterService to be able to unittest

ClusterService does not depend on connection manager
and client anymore.

* Add unittest for ClientClusterServiceImpl

Also revised the fix to cover blue/green case as well.
ClientClusterService will not serve empty memberlist even when
blue/green is happening.