Fix wan address registration issue
With this change, we also register the WAN addresses of the remote
member in the connection initiator side when the connections are
created using the connection manager with EndpointQualifier#MEMBER.

Also, in the previous changes, we were registering the client socket
of the initiator on the acceptor side unnecessarily, removed that
part too. Also, when some member addresses overlap each other in a
cluster topology (A possible case in wan connections in two clusters
using two separate private networks), print a warning log without
overriding the old registered address in this case. In addition, when
connection failures occur during operation invocation, the wrong
connection manager was used to reconnect in some case, fixed that wrong
usage.