Use ConcurrentReferenceHashMap instead of CHM to cache serializers
When we deserialize Externalizable/Serializable we store the
Class->SerializerAdapter in AbstractSerializationService#typeMap.

When the Serializable comes from a Jet job, e.g. a lambda in a Jet
pipeline, the Class instance is different for each instance of the job.
The typeMap keeps the references indefinitely and prevents unloading of
the Jet job's classloader, creating a memory leak.

Also, see https://hazelcast.atlassian.net/browse/HZ-914.

Using ConcurrentReferenceHashMap with weak reference type for the key
allows the class, together with the JetClassloader, to be garbage
collected once the job is finished and the JetClassloader is no longer
strongly referenced.

For the duration of the job the weak reference to the Class will never
be cleared, because there is another strong reference to the class from
the classloader. The same holds for Class objects from Hazelcast's
classloader, so this will not introduce any misses in the `typeMap`
cache.

Tested locally with metaspace limited to 64m with a simple job with a
lambda. On master we get OOM exception before submitting 1000 jobs, with
the fix more than 20 000 jobs. Didn't add a test to the test suite,
because it is covered by Jet soak test suite, where this issue was
originally discovered.