Use dynamicPort for WireMockRule
In #20034 wiremock failed to bind to the port 8080, because it was
already in use. I wasn't able to reproduce it, but it's likely that
these 2 tests interfere with each other, or possibly REST interface in
other test open on port 8080.

This changes the PhoneHomeIntegrationTest and RESTClientPhoneHomeTest,
which use port 8080, to use `dynamicPort()` (random port).

Fixes #20034