[HZ-1140] Added a CPMember UUID check to distinguish a rejoined node and a restarted one.
In addition to comparing IP addresses, we also need to check the UUID of the CP Member to distinguish between a new (or restarted) node and a node that has been rejoined after split-brain healing.

A new `GetLocalCPMemberOp` operation has been added to get a `CPMemberInfo` object from a remote node.