Reset call timeout when rescheduling PromotionCommitOp
The PromotionCommitOperation is executed in 3 stages.
First BeforePromotionOp is invoked, then migration is finalized
(rollback/commitMigration) and finally events are sent.
In between each stage, the same instance of PromotionCommitOperation
(with its original invocation time & calculated call timeout)
is rescheduled for execution in OperationService. If partition threads
are blocked for duration longer than the original call timeout, then
rescheduling fails and an OperationTimeoutException is sent to the
master member. The fix is to reset the invocation time & call timeout
before rescheduling the operation to reflect the fact that the
operation is to be executed anew. The alternative would be to clone
the current operation and schedule the clone, however allocating
more memory for a clone of an existing operation seems wasteful.

Since this is not an operation execution failure (timeout is detected
before the operation is actually called), the operation's
onExecutionFailure method is not called, so there is no opportunity
to react to the particular rescheduling failure.