Generalize TestSupport for exact IO sequence (#21516)
`TestSupport` is used to unit-test processors. However, currently it allows only
supplying of the whole input, and it can assert the whole output. However, unit
tests can test more cases if they could assert that a certain output is produced
after a certain input. This PR adds this possibility: the user can exactly
specify a sequence of input and output events, and the utility asserts the
output after each part of the input. It is also backwards compatible.