Fix SHOW JOBS test
The test fails because it submitted no job, but `testJob` was found. This is
apparently a leftover from other tests which share the cluster and create a job
named `testJob`.

The `SimpleTestInClusterSupport` cancels all jobs after each test, but
`finalizeJob` uses `removeAsync` for deleting the JobRecords. SHOW JOBS used to
read only `JobRecords` map to determine the list of jobs, and it could happen
that the job was cancelled, but the record wasn't yet removed (because the
removal was async), and the next test could still see the record.

The fix is to remove jobs with an existing JobResult from the set.

Fixes #20426