Invalidate client session if CP Subsystem is unaware of it
For FencedLock and Semaphore CP data structures, sessions are required to keep track of the liveliness of callers.
To keep its session alive, a caller commits a periodic heartbeat to the CP group. If the CP Subsystem lost information about this CP group after the restart, there is no way to tell the caller to invalidate the session.

To fix this, we check if a `groupId`, that session belongs to, is known by the CP Subsystem. If it doesn't, it means the CPGroup and session have been destroyed, and we can respond to the caller with `CPGroupDestroyedException`, which invalidates the caller's session.