Simplify WatermarkedFields, fix Union
The `WatermarkedFields` used a `Map<Integer, RexNode>`. The key was the field
index that was watermarked, and the value was the expression used to extract
that value from the rel's input. However, the value is never necessary. If
`rel1` is an input of `rel2`, then `rel2` never needs the expression used by
`rel1` to extract the watermarked value from `rel1`'s input, nor does it have
access to the `rel1`'s input. The `rel1` doesn't need that expression too, as it
must have it by other means, because it creates the output row.

This PR removes the value and replaces `Map<Integer, RexNode>` with just
`Set<Integer>`.

It also fixes the watermarked fields for `Union`: the output of union only
contains the watermark, if _all_ its input have that same field watermarked.
Before it contained it if _any_ input has a column watermarked.