MC-1536 client connection process listener (#21587)
Adds a listener that can be used by the caller to receive fine-grained error events about establishing the connection to the cluster. The purpose of this improvement is being able to provide detailed information for users about why MC cannot connect to a cluster, and where they should look for troubleshooting.

MC-side PR: hazelcast/management-center#5799

JIRA issue: MC-1536

Change details: a new EventListener sub-interface is added which can be set up in a client configuration in the same way as any other existing listener. Despite there is an already existing ConnectionListener, I added a new interface instead of adding more methods to ConnectionListener, because the existing interface is also used in member-side services, and some of these events would be irrelevant there.