Prevent unnecessary retries if the METADATA group is unaware of the requested CPGroup
Information about all CP groups should be present in the METADATA group. If there is no information, there is no point to retry Raft operations.

For example, this behaviour can happen in the case of FencedLock and Semaphore CP data structures. Those data structures require sessions to keep track of the liveliness of callers.
To keep its session alive, a caller commits a periodic heartbeat to the CP group. If the CP Subsystem lost information about this CP group after the restart, there is no way to tell the caller to invalidate the session.

To fix this, we check if CPGroup is known by the METADATA group.
If it is unaware, it means that the CP Subsystem has lost information about CPGroup after the restart, and we can respond with a non-retriable `CPGroupDestroyedException` exception.