[API-1427] Fix race in the TestHazelcastFactory (#21718)
* Fix race in the TestHazelcastFactory

Formerly, we were creating the client instance and after that
registering an instance future for it to the static `CLIENTS` map.

This is fine for the happy path, both under some scenarios, this
implementation is open to dangling client references in the
aforementioned map.

For example, the call to the `client.start()` can return successfully,
but just before we add an instance future to the `CLIENTS` map,
the client might be shutdown by an external thread. That shutdown
process tries to remove the instance future associated with
that client from that map, in
https://github.com/hazelcast/hazelcast/blob/master/hazelcast/src/main/java/com/hazelcast/client/HazelcastClient.java#L367

If the code that adds that future to that map runs after
the code that removes it, we would end up with a dangling
reference to a shutdown client.

To solve this problem, I have refactored
`TestHazelcastFactory#newHazelcastClient` to use the code we already
have, that deals with the lifecycle of the client and the instance
future correctly. This is also good for getting rid of the duplication
in that method.

This was the reason of the recent failures behind https://github.com/hazelcast/hazelcast/issues/15739.

* use unique client names in the `ClusterNameTest` and use `ConcurrentMap`
interface in the field definition