Add time-based expiration to the remembered member addresses
With this commit, TcpIpJoiner clears the remembered addresses of the
previously joined and left members after a certain period of time
passed after they left the cluster. We add a new property called
`PREVIOUSLY_JOINED_MEMBER_ADDRESS_RETENTION_DURATION_PROP` for
specifying this time period with a default value = 4 hours (14400 secs).
When members leave, we record this time in the knownMemberAddresses map,
then we try to clean up this address map in the
TcpIpJoiner#searchForOtherClusters which is called periodically from the
split brain handler.