[API-1425] Do not try to connect cluster members when the client is not connected to cluster (#21705)
* Do not try to connect cluster members when the client is not connected to cluster

We run `ConnectToAllClusterMembersTask` periodically to make sure
the client has a connection to all cluster members, in case
smart routing is enabled.

However, the task was missing a check, and was trying to connect
to the last known member list even though the client was disconnected
from the cluster.

This was problematic in the sense that, it was not respecting the
cluster connection retry configuration of the user. In scenarios
where the user configured backoffs during the reconnection attempts
are long, we were trying to connect back to the cluster every second
anyways.

There is no bullet-proof way to prevent the client from doing so in the
current design, but adding a simple check before each connection attempt
on the task will get rid of almost all such attempts.

The added test could also fail under some unlucky scenarios,
where the task submitted to reconnect the cluster would be delayed
more than 30 seconds. There are (ugly) ways to make this test
reliable by mocking the `WaitStrategy` class but I believe 30 second
wait time would be sufficient for the reconnect task to be
executed. If this test fails in the future, we would choose
the ugly way :)

* introduce a new client state for the disconnected client

* merge two if statements