Add diskUsage and logLength metric for tstore
This PR introduces 2 metrics:

 - logLength
 - diskUsage

Both of these metrics are aggregated per map as this was the
requirement (https://hazelcast.atlassian.net/browse/HZ-1323).

Disk usage increased and decreased when the files actually written to
and deleted from the disk. So this will be accurate indicator of disk
consumption.

Log length is introduced for licensing purposes. This metric is
basically size of valid logic address space. It's increased when a new
allocation occurs and decreased when log is truncated.

Regarding licensing, one improvement could be removing garbage records
from log length. I would advise against that because hybrid log is
unaware of what index does and what records are garbage. This means
that you would need to introduce record garbage mechanisms in
index/storage. One huge disadvantage of this beside its messy
structure is since hybrid log isn't aware of which records are garbage
it must assume all of a file is garbage when truncated. This is
normally not an issue however, if you miss garbage, with time garbage
may even become negative. The places I found in storage that produces
garbage as follows:

 - TieredStorageImpl.removeRecord()
 - TieredStorageImpl.updateRecordValue()
 - TieredStorageImpl.clear()

It's easy to miss one of the places garbage produced, and it's possible
that future changes can be overlooked. This is why I didn't calculate
garbage.

If calculating garbage is a must, although I didn't go into deeper,
I would suggest calculating usage from operations/storage/index in the
future.

Partition based metrics or primary/backup separation is easily doable.
I didn't include them yet because it wasn't a requirement.